Struct slr_config::Error

pub struct Error {
    pub kind: ErrorKind,
    pub text: String,
}

The error type used throughout this crate.

Fields
kind: ErrorKind text: String
Methods

impl Error

pub fn new(kind: ErrorKind, text: String) -> Error

pub fn from_span(
span: Span,
source: Option<&Source>,
kind: ErrorKind,
msg: &str
) -> Error

Creates an error from a certain span of the source. The source argument, if set, must be set to the source that was used when the span was created.