Trait slr_config::ElementRepr

pub trait ElementRepr {
    fn from_element<'l>(&mut self,
                        elem: &ConfigElement,
                        src: Option<&Source<'l>>)
                        -> Result<(), Vec<Error>>;
    fn to_element(&self) -> ConfigElement;
}

Describes a way to convert a type to a ConfigElement and back.

Required Methods

fn from_element<'l>(&mut self,
elem: &ConfigElement,
src: Option<&Source<'l>>)
-> Result<(), Vec<Error>>

Updates the contents of self based on values in the element.

fn to_element(&self) -> ConfigElement

Creates an element that represents the contents of self.